% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CScci.R
\name{add_cellchat_prob}
\alias{add_cellchat_prob}
\title{filter cellchat CCC networks}
\usage{
add_cellchat_prob(conserved_ccc_df, cellchat_df)
}
\arguments{
\item{cellchat_df}{}
}
\description{
Use conserved CCC to filter cellchat CCC networks
}
