% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCMarkerTest.R
\name{AUCMarkerTest}
\alias{AUCMarkerTest}
\title{internal
internal function from seurat}
\usage{
AUCMarkerTest(data1, data2, mygenes, print.bar = TRUE)
}
\arguments{
\item{data1}{cell in cluster 1}

\item{data2}{cell in cluster 2}

\item{mygenes}{gene used}

\item{print.bar}{which apply function}
}
\description{
internal
internal function from seurat
}
