% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDatabase.R
\name{buildHomDatabase}
\alias{buildHomDatabase}
\title{Build homologous gene database}
\usage{
buildHomDatabase(MGI, Species_name1, Species_name2)
}
\arguments{
\item{MGI}{MGI database, download from http://www.informatics.jax.org/}

\item{Species_name1}{The name of the first species. input 'mm' for mouse, 'hs'
for human, 'zf' for zebrafish, 'ch' for chicken, 'cf' for dog, 'pt' for chimpanzee,
'xt' for frog, 'rn' for rat, 'bt' for cattle, and 'rh' for macaque.}

\item{Species_name2}{The name of the second species.  input 'mm' for mouse, 'hs'
for human, 'zf' for zebrafish, 'ch' for chicken, 'cf' for dog, 'pt' for chimpanzee,
'xt' for frog, 'rn' for rat, 'bt' for cattle, and 'rh' for macaque.}
}
\value{
homologous gene database of two species
}
\description{
Build homologous gene database according to Vertebrate Homology
data in MGI database. This function currently supports ten species: cattle,
chicken, chimpanzee, dog, frog, human, macaque, mouse, rat, and zebrafish
}
