% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{CACIMAR_cols}
\alias{CACIMAR_cols}
\title{CACIMAR colors palette}
\usage{
CACIMAR_cols(color_number)
}
\arguments{
\item{color_number}{numeric, indicating used colors number}
}
\value{
vector of colors
}
\description{
CACIMAR colors palette
}
\examples{
CACIMAR_cols(10)
CACIMAR_cols(20)
}
