% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_Conserved_CCI1.R
\name{Caculate_cell_pair_cci_score}
\alias{Caculate_cell_pair_cci_score}
\title{Caculate conserved score of cell-cell interaction}
\usage{
Caculate_cell_pair_cci_score(
  conserved_result_df,
  species1_cci,
  species2_cci,
  conserved_cell_types_df,
  species_name1,
  species_name2
)
}
\arguments{
\item{conserved_result_df}{dataframe, result from function Identify_Conserved_CCI1}

\item{species1_cci}{dataframe, result from SingleCellSignalR of species1}

\item{species2_cci}{dataframe, result from SingleCellSignalR of species2}

\item{conserved_cell_types_df}{dataframe, contain the conserved cell type for each species, like conserved_cell_types_mm_zf <- data.frame("mm" = c("mmRods", "mmRod BC", "mmCones", "mmPericytes", "mmV/E cells", "mmRGC", "mmGABAergic AC", "mmRPE", "mmResting MG", "mmActivated MG", "mmMicroglia"), "zf" = c("zfRods", "zfCone BC", "zfCones", "zfPericytes", "zfV/E cells", "zfRGC", "zfGABAergic AC", "zfRPE", "zfResting MG", "zfActivated MG", "zfMicroglia"))}

\item{species_name1}{two character to represent species1, like "mm". You should set this value from dataframe species_names_ref.rda}

\item{species_name2}{two character to represent species2, like "zf". You should set this value from dataframe species_names_ref.rda}
}
\value{
dataframe, conserved Weights table
}
\description{
caculate the conserved score of cell-cell interaction with summed weights of ligand-receptor interactions for two species
}
\examples{
conserved_cell_types_mm_zf <- data.frame("mm" = c("mmRods", "mmRod BC", "mmCones", "mmPericytes", "mmV/E cells", "mmRGC", "mmGABAergic AC", "mmRPE", "mmResting MG", "mmActivated MG", "mmMicroglia"), "zf" = c("zfRods", "zfCone BC", "zfCones", "zfPericytes", "zfV/E cells", "zfRGC", "zfGABAergic AC", "zfRPE", "zfResting MG", "zfActivated MG", "zfMicroglia"))
cci_conserved_Weights_table_mm_zf <- Caculate_cell_pair_cci_score(conserved_result_df=conserved_result_mm_zf,
species1_cci = SingleCellSignalR_mouse_result,
species2_cci = SingleCellSignalR_zebrafish_result,
conserved_cell_types_df = conserved_cell_types_mm_zf,
species_name1 = "mm",
species_name2 = "zf")
}
