% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_Conserved_CCI2.R
\name{Caculate_cell_pair_cci_score2}
\alias{Caculate_cell_pair_cci_score2}
\title{Caculate conserved score of cell-cell interaction for three species}
\usage{
Caculate_cell_pair_cci_score2(
  conserved_result_df,
  species1_cci,
  species2_cci,
  species3_cci,
  conserved_cell_types_df,
  species_name1,
  species_name2,
  species_name3
)
}
\arguments{
\item{conserved_result_df}{dataframe, result from function Identify_Conserved_CCI2}

\item{species1_cci}{dataframe, result from SingleCellSignalR of species1}

\item{species2_cci}{dataframe, result from SingleCellSignalR of species2}

\item{species3_cci}{dataframe, result from SingleCellSignalR of species3}

\item{conserved_cell_types_df}{dataframe, contain the conserved cell type for each species, like conserved_cell_types_mm_zf_ch <- data.frame('mm' = c("mmResting MG","mmGABAergic AC", "mmRGC", "mmCones"), "zf" = c("zfResting MG", "zfGABAergic AC", "zfRGC", "zfCones"), 'ch' = c("chResting MG","chGABAergic AC", "chRGC", "chCones"))}

\item{species_name1}{two character to represent species1, like "mm". You should set this value from dataframe species_names_ref.rda}

\item{species_name2}{two character to represent species2, like "zf". You should set this value from dataframe species_names_ref.rda}

\item{species_name3}{two character to represent species3, like "ch". You should set this value from dataframe species_names_ref.rda}
}
\value{
dataframe, conserved Weights table
}
\description{
caculate the conserved score of cell-cell interaction with summed weights of ligand-receptor interactions for three species
}
