% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cal_JointPower.R
\name{Cal_JointPower}
\alias{Cal_JointPower}
\title{Calculate the joint power}
\usage{
Cal_JointPower(power01)
}
\arguments{
\item{power01}{power01}
}
\value{
???
}
\description{
Calculate the joint power
}
