% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cal_JointPower2.R
\name{Cal_JointPower2}
\alias{Cal_JointPower2}
\title{Calculate the joint power}
\usage{
Cal_JointPower2(MarkerRoc01)
}
\arguments{
\item{MarkerRoc01}{MarkerRoc01}
}
\value{
???
}
\description{
Calculate the joint power
}
