% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cal_MarkersRoc.R
\name{Cal_MarkersRoc}
\alias{Cal_MarkersRoc}
\title{Calculate the power of each marker gene}
\usage{
Cal_MarkersRoc(pbmc01, marker01)
}
\arguments{
\item{pbmc01}{seurat object}

\item{marker01}{marker gene list}
}
\description{
Calculate the power of each marker gene
}
