% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cal_RelativePower.R
\name{Cal_RelativePower}
\alias{Cal_RelativePower}
\title{Calculate the relative power}
\usage{
Cal_RelativePower(power01)
}
\arguments{
\item{power01}{power}
}
\description{
Calculate the relative power
}
