% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_Cell_interaction_sankey.R
\name{calculate_Weights}
\alias{calculate_Weights}
\title{Sum weight of cell-cell interactions}
\usage{
calculate_Weights(
  species1_cci,
  species2_cci,
  specie_name1 = "Mm",
  specie_name2 = "Zf"
)
}
\arguments{
\item{species1_cci}{data frame,  cell-cell interactions result with perform_CCI_analysis of species 1}

\item{species2_cci}{data frame,  cell-cell interactions result with perform_CCI_analysis of species 2}

\item{specie_name1}{character, name for species 1, like "Mm"}

\item{specie_name2}{character, name for species 2, like "Zf"}
}
\value{
data frame of weight result
}
\description{
Here calculate overall weight of cell-cell interactions within each pair of cell types
}
\examples{
all_weight_df_long <- calculate_Weights(species1_cci = SingleCellSignalR_mouse_result,
species2_cci = SingleCellSignalR_zebrafish_result)

head(all_weight_df_long)
            source       target    weight species scale_weight             Source            Source2
1     Activated MG Activated MG 152.07967      Mm  0.008256281     MmActivated MG     MmActivated MG
2       Astrocytes Activated MG 230.97254      Mm  0.012539310       MmAstrocytes       MmAstrocytes
3            Cones Activated MG  51.89929      Mm  0.002817570            MmCones            MmCones
4     GABAergic AC Activated MG  99.92955      Mm  0.005425093     MmGABAergic AC     MmGABAergic AC
5   Glycinergic AC Activated MG  78.95619      Mm  0.004286467   MmGlycinergic AC   MmGlycinergic AC
6 Horizontal cells Activated MG 106.23687      Mm  0.005767513 MmHorizontal cells MmHorizontal cells
}
