% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChordDiagram.R
\name{ChordDiagram}
\alias{ChordDiagram}
\title{Make a ChordDiagram}
\usage{
ChordDiagram(
  net,
  Score_factor_size = 20,
  brewer_pal_used = "Set3",
  grid.col = NULL,
  grid.border = "#F8F8F8",
  is_link_colors_threshold = TRUE,
  link_colors_threshold = 0.5,
  link_colors_down_threshold_col = "grey",
  filename = "chordDiagram.pdf",
  order_grid = NULL,
  directional = 1,
  direction.type = c("arrows"),
  diffHeight = -0.03,
  annotationTrack = "grid",
  reduce = -1,
  link.arr.type = "big.arrow",
  link.border = "white",
  link.lwd = 1,
  link.sort = TRUE,
  link.decreasing = TRUE,
  link.largest.ontop = TRUE,
  link.overlap = F,
  transparency = 0,
  cex = 1.5,
  ylim_edit = 0,
  facing = "clockwise",
  niceFacing = TRUE,
  col = "black",
  adj = c(0, 0.5),
  start.degree = -5,
  magnify_set = TRUE,
  picture_width = 14,
  picture_height = 10
)
}
\arguments{
\item{net}{data frame, a data frame contain the source, target, and weight}

\item{Score_factor_size}{numeric, scale the weight score for showing}

\item{brewer_pal_used}{character, the brewer_pal in RColorBrewer, like "Set1",  "Set3", "Paired"}

\item{grid.col}{vector, color vectors with names. By default, it is NULL and it will asign colors automatically.}

\item{grid.border}{character, corlors for borders of grids. If it is NULL, the border color is same as grid color}

\item{is_link_colors_threshold}{logical, whether change the colors of links which below a threshold to a  specific color}

\item{link_colors_threshold}{numeric, the threshold used to change colors for links}

\item{link_colors_down_threshold_col}{character, specific corlor for links which below a threshold}

\item{filename}{character, save the plot with this filename}

\item{order_grid}{vector, a ordered vector for grid. By default, it is NULL}

\item{directional}{Directions for the links. 1 means the direction is from the first column in df to the second column, -1 is the reverse, 0 is no direction, and 2 for two directional. The value can be a vector which has same length as number of rows in df}

\item{direction.type}{type for representing directions. Can be one or two values in "diffHeight" and "arrows"}

\item{diffHeight}{numeric, The difference of height between two 'roots' if directional is set to TRUE}

\item{annotationTrack}{annotation for the track. It can set to c("name","grid", "axis"), or one of them.}

\item{reduce}{numeric, if the ratio of the width of certain grid compared to the whole circle is less than this value, the grid is removed on the plot. Set it to value less than zero if you want to keep all tiny grid.}

\item{link.arr.type}{type for the arrow, "big.arrow" is set by default}

\item{link.border}{character, corlors for links}

\item{link.lwd}{numeric, width for link borders}

\item{link.sort}{logical, whether sort links on every sector based on the width of the links on it}

\item{link.decreasing}{logical, for link.sort}

\item{link.largest.ontop}{logical, controls the order of adding links}

\item{link.overlap}{logical, whether the links that come or end in a same sector overlap}

\item{transparency}{numeric, transparency for link colors}

\item{cex}{numeric, font size for the text in annotationTrack}

\item{ylim_edit}{numeric, control the text height from the grid}

\item{facing}{control the direction of the text in annotationTrack, it can be one of c("inside", "outside", "reverse.clockwise", "clockwise", "downward", "bending", "bending.inside", "bending.outside")}

\item{niceFacing}{logical, adjusted the text to fit human eyes}

\item{col}{character, colors for the text in annotationTrack}

\item{adj}{offset for text, like c(0, 0.5). By default the text position adjustment is either horizontal or vertical in the canvas coordinate system.}

\item{start.degree}{numeric, rotation the ChordDiagram with this angle}

\item{picture_width}{numeric, width of the ChordDiagram for saving}

\item{picture_height}{numeric, height of the ChordDiagram for saving}
}
\value{
saving ChordDiagram in current directory
}
\description{
Construct a ChordDiagram to show the conservation score for intercellular interactions
}
\examples{
load(system.file("extdata", "CCC_conserved_sumary.rda", package = "CACIMAR"))
cci_data = CCC_conserved_sumary[, c("Source", "Target", "score_weight")]
ChordDiagram(net = cci_data, filename = "chordDiagram.pdf", link_colors_threshold = 0.75)

}
