% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_Conserved_CCI2.R
\name{conserved_interaction_score}
\alias{conserved_interaction_score}
\title{Caculate conserved interaction score and Make data for ChordDiagram}
\usage{
conserved_interaction_score(
  conserved_result_species,
  SingleCellSignalR_sp1_result,
  SingleCellSignalR_sp2_result,
  SingleCellSignalR_sp3_result,
  conserved_cell_types_df,
  species_name1 = "mm",
  species_name2 = "zf",
  species_name3 = "ch"
)
}
\arguments{
\item{conserved_result_species}{dataframe, result from function Identify_Conserved_CCI2}

\item{SingleCellSignalR_sp1_result}{dataframe, result from SingleCellSignalR of species1}

\item{SingleCellSignalR_sp2_result}{dataframe, result from SingleCellSignalR of species2}

\item{SingleCellSignalR_sp3_result}{dataframe, result from SingleCellSignalR of species3}

\item{conserved_cell_types_df}{dataframe, contain the conserved cell type for each species, like conserved_cell_types_mm_zf_ch <- data.frame('mm' = c("mmResting MG","mmGABAergic AC", "mmRGC", "mmCones"), "zf" = c("zfResting MG", "zfGABAergic AC", "zfRGC", "zfCones"), 'ch' = c("chResting MG","chGABAergic AC", "chRGC", "chCones"))}

\item{species_name1}{two character to represent species1, like "mm". You should set this value from dataframe species_names_ref.rda}

\item{species_name2}{two character to represent species2, like "zf". You should set this value from dataframe species_names_ref.rda}

\item{species_name3}{two character to represent species3, like "ch". You should set this value from dataframe species_names_ref.rda}
}
\value{
list of conserved Weights table and data for chordDiagram
}
\description{
Caculate conserved interaction score and Make data for ChordDiagram
}
\examples{
#not run
#conserved_result_mm_zf_ch is the result from function Identify_Conserved_CCI2
conserved_cci_result <- conserved_interaction_score(conserved_result_species = conserved_result_mm_zf_ch,
SingleCellSignalR_sp1_result = SingleCellSignalR_mouse_result,
SingleCellSignalR_sp2_result = SingleCellSignalR_zebrafish_result,
SingleCellSignalR_sp3_result = SingleCellSignalR_chick_result,
conserved_cell_types_df = conserved_cell_types_mm_zf_ch,
species_name1 = "mm",
species_name2 = "zf",
species_name3 = "ch")
}
