% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Converse_GeneIDSymbol.R
\name{Converse_GeneIDSymbol}
\alias{Converse_GeneIDSymbol}
\title{Inner function to make ID change}
\usage{
Converse_GeneIDSymbol(Gene1, GeneInf1 = NULL)
}
\description{
Inner function to make ID change
}
