% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{FormatConservedMarkers}
\alias{FormatConservedMarkers}
\title{Format Conserved markers to plot heatmap}
\usage{
FormatConservedMarkers(ConservedMarker)
}
\arguments{
\item{ConservedMarker}{Result from 'Identify_ConservedMarkers'}
}
\value{
list contains two data.frame
}
\description{
Format Conserved markers to plot heatmap
}
\examples{
load(system.file("extdata", "zf_mm_markers.rda", package = "CACIMAR"))
ConservedMarker <- Identify_ConservedMarkers(OrthG_Mm_Zf,Mm_marker,Zf_marker,
Species_name1 = 'mm',Species_name2 = 'zf')
MarkersPlot <- FormatConservedMarkers(ConservedMarker)
}
