% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Used_OrthG.R
\name{Get_OrthG}
\alias{Get_OrthG}
\title{Inner function to get orthology}
\usage{
Get_OrthG(OrthG1, MmRNA1, ZfRNA1, Spec1, MmPattern1 = "", ZfPattern1 = "")
}
\description{
Inner function to get orthology
}
