% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refine_markers.R
\name{Get_scRNA_AggExp}
\alias{Get_scRNA_AggExp}
\title{Inner function to refine markers}
\usage{
Get_scRNA_AggExp(
  pbmc,
  Symb1,
  Cluster1 = c("all", "res.0.1"),
  Group1 = c("all", "protocol"),
  ExpType1 = c("expression", "fraction")
)
}
\description{
Inner function to refine markers
}
