% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-species_markers.R
\name{Get_Used_OrthG}
\alias{Get_Used_OrthG}
\title{Identify orthologs for two species according to power}
\usage{
Get_Used_OrthG(OrthG1, Species1_expression, Species2_expression, Species)
}
\arguments{
\item{Species}{}
}
\value{

}
\description{
Identify orthologs for two species according to power
}
