% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNAseq_species.R
\name{get_Wilcox_Markers_Cond}
\alias{get_Wilcox_Markers_Cond}
\title{Inner function to get wilcox markers cond}
\usage{
get_Wilcox_Markers_Cond(
  Marker1,
  Cor1,
  CorThr1 = 0.2,
  CorType1 = "Cor_mmNMDA_mmLD"
)
}
\description{
Inner function to get wilcox markers cond
}
