% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_Cor.R
\name{Heatmap_Cor}
\alias{Heatmap_Cor}
\title{plot the pheatmap of marker genes across different species}
\usage{
Heatmap_Cor(
  RNA1,
  CellType,
  RowType1 = "",
  ColType1 = "",
  cluster_cols = F,
  cluster_rows = F,
  Color1 = NULL
)
}
\arguments{
\item{RNA1}{expression}

\item{CellType}{cell type}

\item{RowType1}{??}

\item{ColType1}{??}

\item{cluster_cols}{boolean values determining if columns should be clustered or hclust object}

\item{cluster_rows}{boolean values determining if rows should be clustered or hclust object}

\item{Color1}{vector of colors used in heatmap}
}
\value{
return a list used in Plot_tree
}
\description{
plot the pheatmap of marker genes across different species
}
