% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_CellType_evolution.R
\name{Identify_CellType_evolution}
\alias{Identify_CellType_evolution}
\title{Title Identify celltype evolution between two species}
\usage{
Identify_CellType_evolution(
  species1_seurat,
  species2_seurat,
  ConservedMarker,
  species1_name = "Mm",
  species2_name = "Zf",
  Used_species1_ID,
  Used_species2_ID,
  dist.method = "euclidean",
  hclust.method = "average",
  layout.tree = "rectangular",
  geom_nodepoint = 0,
  col.value = c("#990000", "#660099")
)
}
\arguments{
\item{species1_seurat}{a species1 seurat object. The seurat object input should have been normalized, and active.ident should be the celltype.}

\item{species2_seurat}{a species2 seurat object. The seurat object input should have been normalized, and active.ident should be the celltype.}

\item{ConservedMarker}{A marker table with homologous genes of the species you input. The homologous genes should all be in your seurat object, respectively. Or you can used the result of Identify_ConservedMarkers function.}

\item{species1_name}{character, indicating the species names of species1.}

\item{species2_name}{character, indicating the species names of species2.}

\item{Used_species1_ID}{character, the columan name in ConservedMarker table for species1 seurat object. This must be in the column names of ConservedMarker.}

\item{Used_species2_ID}{character, the columan name in ConservedMarker table for species2 seurat object. This must be in the column name of ConservedMarker.}

\item{dist.method}{The method used for caculating distance. It can be any one used in dist funtion of R package stats.}

\item{hclust.method}{The method used for hierarchical cluster analysis. This can be one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).}

\item{layout.tree}{The layout style of the tree. It can be one of 'rectangular', 'dendrogram', 'slanted', 'ellipse', 'roundrect', 'fan', 'circular', 'inward_circular', 'radial', 'equal_angle', 'daylight' or 'ape'.}

\item{geom_nodepoint}{numeric, if it is greater than 0, it will show the nodepoint in the tree.}

\item{col.value}{a vector for the colors used for the tippoints and tiplabels. The colors should be a length of the number of species, here that is tow.}
}
\value{
a list, it contains the tree result, and a plot of the tree.
}
\description{
This function uses conserved markers between two species to find the correlation of celltypes between species.
}
\examples{
Identify_CellType_evolution(Zf_seurat, Mm_seurat, ConservedMarker, species1_name = "Zf", species2_name = "Mm", Used_species1_ID = "Used_zf_ID", Used_species2_ID = "Used_mm_ID", dist.method = "euclidean", hclust.method = "average", layout.tree = 'rectangular', geom_nodepoint = 0, col.value = c("#990000", "#660099"))
}
