% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_CellTypes1.R
\name{Identify_CellTypes1}
\alias{Identify_CellTypes1}
\title{caculate ROC}
\usage{
Identify_CellTypes1(pbmc, Marker1)
}
\arguments{
\item{pbmc}{seurat object}

\item{Marker1}{marker gene list}
}
\value{
???
}
\description{
caculate ROC
}
