% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_CellTypes2.R
\name{Identify_CellTypes2}
\alias{Identify_CellTypes2}
\title{Calculate the relative power of cell types for each cell cluster}
\usage{
Identify_CellTypes2(MarkerRoc1)
}
\arguments{
\item{MarkerRoc1}{MarkerRoc1}
}
\value{
???
}
\description{
Calculate the relative power of cell types for each cell cluster
}
