% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CScci.R
\name{Identify_Conserved_CCI}
\alias{Identify_Conserved_CCI}
\title{Title}
\usage{
Identify_Conserved_CCI(OrthG, sp1_ccc, sp2_ccc, Species_name1, Species_name2)
}
\arguments{
\item{OrthG}{ortholog genes database}

\item{sp1_ccc}{cell-cell interactions in species1. First column should
be ligand, second column should be target, third column should be corresponding
cell type of ligand, fourth column should be corresponding cel type of target,
fifth column should be weight of ligand-receptor interaction}

\item{sp2_ccc}{cell-cell interactions in species2. First column should
be ligand, second column should be target, third column should be corresponding
cell type of ligand, fourth column should be corresponding cel type of target,
fifth column should be weight of ligand-receptor interaction}

\item{Species_name1}{character, indicating the species names of Species1_GRN}

\item{Species_name2}{character, indicating the species names of Species2_GRN}
}
\description{
Title
}
