% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_Conserved_CCI1.R
\name{Identify_Conserved_CCI1}
\alias{Identify_Conserved_CCI1}
\title{Identify conserved cell-cell interaction from two species}
\usage{
Identify_Conserved_CCI1(
  species1_cci,
  species2_cci,
  species_names_ref = NULL,
  conserved_cell_types_df,
  species_name1 = "mm",
  species_name2 = "zf",
  HOM_matrix = NULL
)
}
\arguments{
\item{species1_cci}{result of SingCellSignalR of species1}

\item{species2_cci}{result of SingCellSignalR of species2}

\item{species_names_ref}{dataframe, if you construct new homolog database and use new species homolog information in this analysis, you should provide the names corresponding relationship. it should be a dataframe contain at lease two column full_name and sy_name.}

\item{conserved_cell_types_df}{dataframe, like conserved_cell_types_mm_zf <- data.frame("mm" = c("mmRods", "mmRod BC", "mmCones", "mmPericytes", "mmV/E cells", "mmRGC", "mmGABAergic AC", "mmRPE", "mmResting MG", "mmActivated MG", "mmMicroglia"), "zf" = c("zfRods", "zfCone BC", "zfCones", "zfPericytes", "zfV/E cells", "zfRGC", "zfGABAergic AC", "zfRPE", "zfResting MG", "zfActivated MG", "zfMicroglia"))}

\item{species_name1}{two character to represent species1, like "mm". You should set this value from dataframe species_names_ref.rda}

\item{species_name2}{two character to represent species2, like "zf". You should set this value from dataframe species_names_ref.rda}

\item{HOM_matrix}{dataframe, default is NULL, if you used new created homolog database, you should put it here}
}
\value{
list for conserved result for species1 and species2
}
\description{
get conserved cell-cell interaction between two species
}
\examples{
conserved_cell_types_mm_zf <- data.frame("mm" = c("mmRods", "mmRod BC", "mmCones", "mmPericytes", "mmV/E cells", "mmRGC", "mmGABAergic AC", "mmRPE", "mmResting MG", "mmActivated MG", "mmMicroglia"), "zf" = c("zfRods", "zfCone BC", "zfCones", "zfPericytes", "zfV/E cells", "zfRGC", "zfGABAergic AC", "zfRPE", "zfResting MG", "zfActivated MG", "zfMicroglia"))
conserved_result_mm_zf <- Identify_Conserved_CCI1(species1_cci=SingleCellSignalR_mouse_result,
                                                  species2_cci=SingleCellSignalR_zebrafish_result,
                                                  conserved_cell_types_df=conserved_cell_types_mm_zf,
                                                  species_name1 = "mm",
                                                  species_name2 = "zf")
}
