% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_Conserved_CCI2.R
\name{Identify_Conserved_CCI2}
\alias{Identify_Conserved_CCI2}
\title{Identify conserved cell-cell interaction from three species}
\usage{
Identify_Conserved_CCI2(
  species1_cci,
  species2_cci,
  species3_cci,
  species_names_ref = NULL,
  conserved_cell_types_df,
  species_name1 = "mm",
  species_name2 = "zf",
  species_name3 = "ch",
  HOM_matrix = NULL
)
}
\arguments{
\item{species1_cci}{result of SingCellSignalR of species1}

\item{species2_cci}{result of SingCellSignalR of species2}

\item{species3_cci}{result of SingCellSignalR of species3}

\item{species_names_ref}{dataframe, if you construct new homolog database and use new species homolog information in this analysis, you should provide the names corresponding relationship. it should be a dataframe contain at lease two column full_name and sy_name.}

\item{conserved_cell_types_df}{dataframe, like conserved_cell_types_mm_zf_ch <- data.frame('mm' = c("mmResting MG","mmGABAergic AC", "mmRGC", "mmCones"), "zf" = c("zfResting MG", "zfGABAergic AC", "zfRGC", "zfCones"), 'ch' = c("chResting MG","chGABAergic AC", "chRGC", "chCones"))}

\item{species_name1}{two character to represent species1, like "mm". You should set this value from dataframe species_names_ref.rda}

\item{species_name2}{two character to represent species2, like "zf". You should set this value from dataframe species_names_ref.rda}

\item{species_name3}{two character to represent species3, like "ch". You should set this value from dataframe species_names_ref.rda}

\item{HOM_matrix}{dataframe, default is NULL, if you used new created homolog database, you should put it here}
}
\value{
list for conserved result for each species
}
\description{
find the conserved cell-cell interaction from three species and caculate the conserved score of cell-cell interaction
}
