% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-species_markers.R
\name{identify_conserved_gene}
\alias{identify_conserved_gene}
\title{Identify conserved gene}
\usage{
identify_conserved_gene(
  OrthG,
  spc1_marker,
  spc2_marker,
  Species_name1,
  Species_name2
)
}
\arguments{
\item{OrthG}{ortholog genes database}

\item{spc1_marker}{vector, indicating the gene of species 1}

\item{spc2_marker}{vector, indicating the gene of species 2}

\item{Species_name1}{character, indicating the species names of Species1_Marker_table.}

\item{Species_name2}{character, indicating the species names of Species2_Marker_table}
}
\description{
Identify conserved gene
}
