% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-species_markers.R
\name{identify_conserved_marker}
\alias{identify_conserved_marker}
\title{Identify conserved markers in conserved celltype}
\usage{
identify_conserved_marker(
  OrthG,
  Species1_Marker_table,
  Species2_Marker_table,
  Species_name1,
  Species_name2,
  conserved_celltype_pair
)
}
\arguments{
\item{OrthG}{ortholog genes database}

\item{Species1_Marker_table}{data.frame of species 1, should contain 'gene' and
'Allcluster' columns.}

\item{Species2_Marker_table}{data.frame of species 2, should contain 'gene' and
'Allcluster' columns.}

\item{Species_name1}{character, indicating the species names of Species1_Marker_table.}

\item{Species_name2}{character, indicating the species names of Species2_Marker_table}

\item{conserved_celltype_pair}{character, indicating the conserved celltypes}
}
\description{
Identify orthologs marker genes for two species based on orthologs database
}
\details{
Identify orthologs marker genes for two species
}
