% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypeFraction.R
\name{identify_conserved_pair}
\alias{identify_conserved_pair}
\title{Identify conserved pair}
\usage{
identify_conserved_pair(conserved_score, quantile_threshold = 0.75)
}
\arguments{
\item{conserved_score}{conserved celltype/GRN score}

\item{quantile_threshold}{numeric, indicating}
}
\description{
Identify conserved pair
}
