% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CScci.R
\name{Identify_ConservedCCI}
\alias{Identify_ConservedCCI}
\title{Identify conserved cell-cell interactions in conserved cell types}
\usage{
Identify_ConservedCCI(
  OrthG,
  Species1_CCI,
  Species2_CCI,
  ConservedCellType = NULL,
  Species_name1,
  Species_name2
)
}
\arguments{
\item{OrthG}{ortholog genes database}

\item{Species1_CCI}{cell-cell interactions in species1. First column should
be ligand, second column should be target, third column should be corresponding
cell type of ligand, fourth column should be corresponding cel type of target}

\item{Species2_CCI}{cell-cell interactions in species2. First column should
be ligand, second column should be target, third column should be corresponding
cell type of ligand, fourth column should be corresponding cel type of target}

\item{Species_name1}{character, indicating the species names of Species1_GRN}

\item{Species_name2}{character, indicating the species names of Species2_GRN}
}
\description{
Identify conserved cell-cell interactions in conserved cell types
}
\examples{
load(system.file("extdata", "cci_test.rda", package = "CACIMAR"))
Identify_ConservedCCI(OrthG_Hs_Mm,hs_cci_test,mm_cci_test,celltype,'hs','mm')
}
