% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSNetworks.R
\name{identify_ct_ConservedNetworks}
\alias{identify_ct_ConservedNetworks}
\title{Identify conserved cell type specific regulatory networks}
\usage{
identify_ct_ConservedNetworks(
  OrthG,
  Species1_GRN,
  Species2_GRN,
  Species_name1,
  Species_name2,
  network_regulation_num = NULL
)
}
\arguments{
\item{OrthG}{ortholog genes database}

\item{Species1_GRN}{gene regulatory network of species 1}

\item{Species2_GRN}{gene regulatory network of species 2}

\item{Species_name1}{character, indicating the species names of Species1_GRN}

\item{Species_name2}{character, indicating the species names of Species2_GRN}
}
\description{
Identify conserved cell type specific regulatory networks
}
