% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_Markers.R
\name{Identify_Markers}
\alias{Identify_Markers}
\title{Identify markers of each cluster}
\usage{
Identify_Markers(
  Seurat_object,
  PowerThr1 = 0.4,
  Spec1 = NULL,
  GeneSymb1 = NULL,
  FracThr1 = 3
)
}
\arguments{
\item{Seurat_object}{Seurat object, should contain cluster information}

\item{PowerThr1}{numeric, indicating}

\item{Spec1}{character, When GeneSymb1 == NULL, inner file will be used for iD changing, this parameter indicate the species of}

\item{GeneSymb1}{Gene correspondence that is used for ID chaning. first column should be ENSEMBLE ID, second column should be Symbol, third column should be NCBIID and fourth column should be officalsymbol.}

\item{FracThr1}{numeric, indicating}
}
\value{

}
\description{
Identify markers of each cluster
}
