% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_Markers1.R
\name{Identify_Markers1}
\alias{Identify_Markers1}
\title{Inner function for markers identifying}
\usage{
Identify_Markers1(Seurat_object, PowerThr1 = 1/3)
}
\description{
Inner function for markers identifying
}
