% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_Markers2.R
\name{Identify_Markers2}
\alias{Identify_Markers2}
\title{Inner function for indentifying markers}
\usage{
Identify_Markers2(pbmc, Marker, GeneSymb1 = NULL, PowerThr1 = 1/3)
}
\description{
Inner function for indentifying markers
}
