% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refine_markers.R
\name{Identify_Markers3}
\alias{Identify_Markers3}
\title{Inner function to refine markers}
\usage{
Identify_Markers3(pbmc, MarkerRoc2, Cluster1, Symb1, FracThr1 = 5)
}
\description{
Inner function to refine markers
}
