% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSNetworks.R
\name{identify_network_relationships}
\alias{identify_network_relationships}
\title{Identify conserved node and edge in the networks
Based on the Conservation analysis result from Identify_ConservedNetworks,
this function further identify the conserved node and edge in the network(graph)}
\usage{
identify_network_relationships(
  Species1_GRN,
  Species2_GRN,
  ConservedNetworkTable,
  Species1_group = "",
  Species2_group = ""
)
}
\arguments{
\item{Species1_GRN}{gene regulatory network of species 1}

\item{Species2_GRN}{gene regulatory network of species 2}

\item{ConservedNetworkTable}{result from Identify_ConservedNetworks}

\item{Species1_group}{character,indicating interested GRN group in species 1}

\item{Species2_group}{character,indicating interested GRN group in species 2}
}
\description{
Identify conserved node and edge in the networks
Based on the Conservation analysis result from Identify_ConservedNetworks,
this function further identify the conserved node and edge in the network(graph)
}
