% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypeFraction.R
\name{Identify_SharedMarkers}
\alias{Identify_SharedMarkers}
\title{Calculate fraction of shared markers between species}
\usage{
Identify_SharedMarkers(Marker_list, Species_names, PowerTh1 = 0.1)
}
\arguments{
\item{Marker_list}{each element should be marker genes table}

\item{Species_names}{character, indicating the species of each marker genes
table in Marker_list}

\item{PowerTh1}{numeric, indicating the threshold to filter marker genes}
}
\value{

}
\description{
Calculate fraction of shared markers between species
}
