% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identy_CellType.R
\name{Identy_CellType}
\alias{Identy_CellType}
\title{Title}
\usage{
Identy_CellType(object, Marker1)
}
\arguments{
\item{object}{seurat object}

\item{Marker1}{marker gene list rownames is symbol first column is ENS ID second column is cell type}
}
\value{
a lsit with cell type
}
\description{
Title
}
