% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MarkersHeatmap.R
\name{Kmeans_heatmap}
\alias{Kmeans_heatmap}
\title{inner function to make heatmap}
\usage{
Kmeans_heatmap(
  RNA1,
  K1 = 1,
  Gene1 = NULL,
  RowGroup1 = NULL,
  ColumnGroup1 = NULL,
  Scale1 = "none",
  Range1 = c(-Inf, Inf),
  Reorder1 = T,
  RevOrder1 = -1,
  clustering_distance_rows = "euclidean",
  clustering_method = "complete",
  Show.Module = F,
  Color1 = NULL,
  show_colnames = T,
  NumRowBlank1 = 10,
  NumColumnBlank1 = 10,
  NAcolum1 = NULL,
  ModuleColor1 = NULL,
  ModuleScale1 = 10,
  cluster_cols = F,
  fontsize = 10,
  legend1 = T,
  border_color = "white",
  ByPanel = F
)
}
\description{
inner function to make heatmap
}
