% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_Conserved_CCI1.R
\name{Make_ChordDiagram_data}
\alias{Make_ChordDiagram_data}
\title{Make data fit for ChordDiagram}
\usage{
Make_ChordDiagram_data(
  cci_conserved_Weights_table,
  species_name1,
  species_name2
)
}
\arguments{
\item{cci_conserved_Weights_table}{dataframe, result from function Caculate_cell_pair_cci_score}

\item{species_name1}{two character to represent species1, like "mm"}

\item{species_name2}{two character to represent species1, like "zf"}
}
\value{
dataframe, data for ChordDiagram
}
\description{
Make data fit for ChordDiagram
}
\examples{
cci_data_mm_zf <- Make_ChordDiagram_data(cci_conserved_Weights_table = cci_conserved_Weights_table_mm_zf, species_name1 = "mm", species_name2 = "zf")
}
