% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_Conserved_CCI2.R
\name{Make_ChordDiagram_data2}
\alias{Make_ChordDiagram_data2}
\title{Make data fit for ChordDiagram for conserved cell types from three species}
\usage{
Make_ChordDiagram_data2(
  cci_conserved_Weights_table,
  species_name1,
  species_name2,
  species_name3
)
}
\arguments{
\item{cci_conserved_Weights_table}{dataframe, result from function Caculate_cell_pair_cci_score2}

\item{species_name1}{two character to represent species1, like "mm"}

\item{species_name2}{two character to represent species2, like "zf"}

\item{species_name3}{two character to represent species3, like "ch"}
}
\value{
dataframe, data for ChordDiagram
}
\description{
Make data fit for ChordDiagram for conserved cell types from three species
}
\examples{
# cci_conserved_Weights_table_sp1_sp2_sp3 is the result of function Caculate_cell_pair_cci_score2
 cci_data_sp1_sp2_sp3 <- Make_ChordDiagram_data2(cci_conserved_Weights_table = cci_conserved_Weights_table_sp1_sp2_sp3,
 species_name1 = "mm",
 species_name2 = "zf",
 species_name3 = "ch")
}
