% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheatmap_LR_multi.R
\name{make_pheatmap_LR_data}
\alias{make_pheatmap_LR_data}
\title{Make ligand-receptor data for pheatmap}
\usage{
make_pheatmap_LR_data(
  cci_conserved_results,
  seurat_object_sp1,
  seurat_object_sp2,
  seurat_object_sp3,
  species_name1 = "mm",
  species_name2 = "zf",
  species_name3 = "ch",
  avg_group = "cellname",
  cci_conserved_Weights,
  subset_quantile = 0.75,
  assay_set = "RNA"
)
}
\arguments{
\item{cci_conserved_results}{dataframe, the result from function Identify_Conserved_CCI2}

\item{seurat_object_sp1}{seurat object of species1}

\item{seurat_object_sp2}{seurat object of species2}

\item{seurat_object_sp3}{seurat object of species3}

\item{species_name1}{character, two character the representation the species1, like "mm"}

\item{species_name2}{character, two character the representation the species2, like "zf"}

\item{species_name3}{character, two character the representation the species2, like "ch"}

\item{avg_group}{character, which group used to calculate the average expression, like "celltype" in metadata}

\item{cci_conserved_Weights}{dataframe, the result from function Caculate_cell_pair_cci_score}

\item{subset_quantile}{numeric, the value to filter the conserved score of the cell-cell interaction, default is 0.75 of the consered score of cell-cell interaction}

\item{assay_set}{which assay used for average expression, like "RNA", or "SCT"}
}
\value{
list, contains the dataframe for heatmap, and the average expression for each species
}
\description{
prepared average expression data for pheatmap
}
\examples{
merge_avg_width_df <- make_pheatmap_LR_data(cci_conserved_results = conserved_result_mm_zf_ch,
seurat_object_sp1 = Mm_seurat,
seurat_object_sp2 = Zf_seurat,
seurat_object_sp3 = ch_seurat,
avg_group = "cellname",
species_name1 = "mm",
species_name2 = "zf",
species_name3 = "ch",
cci_conserved_Weights = conserved_cci_result$cci_conserved_Weights)
}
