% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Markertest.R
\name{Markertest}
\alias{Markertest}
\title{internal function from seurat2.3.4 to test marker gene}
\usage{
Markertest(
  object,
  cells.1,
  cells.2,
  genes.use = NULL,
  print.bar = TRUE,
  assay.type = "RNA"
)
}
\arguments{
\item{object}{seurat object}

\item{cells.1}{cell in cluster 1}

\item{cells.2}{cell in cluster 2}

\item{genes.use}{gene to be tested}

\item{print.bar}{use which apply function}

\item{assay.type}{???}
}
\value{
Returns a 'predictive power' (abs(AUC-0.5) * 2) ranked matrix of
}
\description{
internal function from seurat2.3.4 to test marker gene
}
