% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Average_expression_heatmap.R
\name{Merge_data_for_heatmap}
\alias{Merge_data_for_heatmap}
\title{Merge the average expression dataframe}
\usage{
Merge_data_for_heatmap(LRpair_show_df1, LRpair_show_df2)
}
\arguments{
\item{LRpair_show_df1}{LRpair_show_df1 must contain columns: Cell_Pair, species, ligand_receptor_avg. Besides, check and make column species only contain two characters}

\item{LRpair_show_df2}{LRpair_show_df2 must contain columns: Cell_Pair, species, ligand_receptor_avg, Besides, check and make column species only contain two characters}
}
\description{
This function merge the average expression dataframe of two species generated by function get_average_expression
}
\examples{
merge_width_df <- Merge_data_for_heatmap(LRpair_show_Mm, LRpair_show_Zf)
}
