% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{OrthG_Hs_Ch}
\alias{OrthG_Hs_Ch}
\title{Orthologs genes database for homo sapiens and zebrafish}
\format{
An object of class \code{data.frame} with 12219 rows and 5 columns.
}
\usage{
OrthG_Hs_Ch
}
\description{
Orthologs genes database for homo sapiens and zebrafish
}
\keyword{datasets}
