% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{OrthG_Hs_Mm}
\alias{OrthG_Hs_Mm}
\title{Orthologs genes database for homo sapiens and mus musculus}
\format{
An object of class \code{data.frame} with 16754 rows and 5 columns.
}
\usage{
OrthG_Hs_Mm
}
\description{
Orthologs genes database for homo sapiens and mus musculus
}
\keyword{datasets}
