% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{OrthG_Hs_Zf}
\alias{OrthG_Hs_Zf}
\title{Orthologs genes database for homo sapiens and zebrafish}
\format{
An object of class \code{data.frame} with 12017 rows and 5 columns.
}
\usage{
OrthG_Hs_Zf
}
\description{
Orthologs genes database for homo sapiens and zebrafish
}
\keyword{datasets}
