% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{OrthG_Mm_Ch}
\alias{OrthG_Mm_Ch}
\title{Orthologs genes database for mus musculus and chicken}
\format{
An object of class \code{data.frame} with 62661 rows and 5 columns.
}
\usage{
OrthG_Mm_Ch
}
\description{
Orthologs genes database for mus musculus and chicken
}
\keyword{datasets}
