% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{OrthG_Mm_Zf}
\alias{OrthG_Mm_Zf}
\title{Orthologs genes database for mus musculus and zebrafish}
\format{
An object of class \code{data.frame} with 65631 rows and 5 columns.
}
\usage{
OrthG_Mm_Zf
}
\description{
Orthologs genes database for mus musculus and zebrafish
}
\keyword{datasets}
