% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-species_markers.R
\name{OrthG_ThreeSpecies}
\alias{OrthG_ThreeSpecies}
\title{Identify orthologs genes for three species}
\usage{
OrthG_ThreeSpecies(
  OrthG_TwoSpecies1,
  OrthG_TwoSpecies2,
  Species_name,
  match_cell_name = NULL
)
}
\arguments{
\item{OrthG_TwoSpecies1}{First two species orthologs marker genes,
generated by OrthG_TwoSpecies. One species in OrthG_TwoSpecies1
should overlap with one species in OrthG_TwoSpecies2}

\item{OrthG_TwoSpecies2}{Second two species orthologs marker genes,
generated by OrthG_Twospecies. One species in OrthG_TwoSpecies2
should overlap with one species in OrthG_TwoSpecies1}

\item{Species_name}{character, indicating the names of species in OrthG_TwoSpecies1
and OrthG_TwoSpecies2}

\item{match_cell_name}{characters contained in both cell names
to match similar cell types}
}
\value{
Data frame of conserved markers
}
\description{
Identify orthologs genes for three species based on the result
of OrthG_Twospecies
}
\examples{
load(system.file("extdata", "CellMarkers.rda", package = "CACIMAR"))
OrthG_TwoSpecies1 <- Identify_ConservedMarkers(OrthG_Mm_Zf,Mm_marker_cell_type
,Zf_marker_cell_type,Species_name1 = 'mm',Species_name2 = 'zf')
OrthG_TwoSpecies2 <- Identify_ConservedMarkers(OrthG_Zf_Ch,Ch_marker_cell_type
,Zf_marker_cell_type,Species_name1 = 'ch',Species_name2 = 'zf')
OrthG_ThreeSpecies<-OrthG_ThreeSpecies(OrthG_TwoSpecies1,OrthG_TwoSpecies2,c('mm','zf','zf','ch'))
}
