% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{OrthG_Zf_Ch}
\alias{OrthG_Zf_Ch}
\title{Orthologs genes database for mus zebrafish and chicken}
\format{
An object of class \code{data.frame} with 38394 rows and 5 columns.
}
\usage{
OrthG_Zf_Ch
}
\description{
Orthologs genes database for mus zebrafish and chicken
}
\keyword{datasets}
