% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-species_markers.R
\name{Overlap_Markers_Cond}
\alias{Overlap_Markers_Cond}
\title{Overlap markers from different conditions of the same species}
\usage{
Overlap_Markers_Cond(markers_expression, celltype, Species)
}
\arguments{
\item{markers_expression}{marker genes information with expression}

\item{celltype}{cell type data.frame, first column should be samples, second
column should be cluster, thrid column should be cell type}

\item{Species}{character, indicating the species of your data}
}
\value{
marker genes information in different samples
}
\description{
Overlap markers from different conditions of the same species
}
