% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_Cell_interaction_sankey.R
\name{perform_CCI_analysis}
\alias{perform_CCI_analysis}
\title{Cell-cell interaction analysis with SingleCellSignalR algorithm}
\usage{
perform_CCI_analysis(
  seurat_obj,
  expre_cell = 10,
  select_DB = "Consensus",
  target_organism,
  method = "sca",
  LRscore_threshold = 0.5,
  scale_score = TRUE
)
}
\arguments{
\item{seurat_obj}{seurat object, a seurat object with cell types in active idents}

\item{expre_cell}{numeric, filter genes expressed in more than expre_cell cells}

\item{select_DB}{character, databased used for analysis, more option can run liana::show_resources()}

\item{target_organism}{ncbi_taxid' or 'name' of the target organism. See ‘show_homologene' for available organisms via OmnipathR’s 'HomoloGene'}

\item{method}{character, method use for cell-cell interaction analysis. By default, we set it to SingleCellSignalR}

\item{LRscore_threshold}{numeric, the threshold the filter the cell-cell interactions, the larger the lesser interactions left}

\item{scale_score}{logical, whether scale the score of the interaction,default is TRUE}
}
\value{
data frame of cell cell interaction analysis result
}
\description{
Perform cell-cell interaction with SingleCellSignalR algorithm contained in liana package.
}
\examples{
# Mouse analysis
SingleCellSignalR_mouse_result <- perform_CCI_analysis(seurat_obj=Mm_seurat, target_organism=10090)

SingleCellSignalR_zebrafish_result <- perform_CCI_analysis(seurat_obj=Zf_seurat, target_organism=7955)
}
