% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Average_expression_heatmap.R
\name{pheatmap_LR}
\alias{pheatmap_LR}
\title{Plot heatmap for average exprssion of ligand and receptor pair}
\usage{
pheatmap_LR(
  width_df,
  filename = "Pheatmap_LR.pdf",
  color_pheatmap_set = NULL,
  colors_list = NULL,
  annotation_names_col = TRUE,
  border_color = "white",
  scale = "none",
  cluster_rows = FALSE,
  cluster_cols = FALSE,
  legend = TRUE,
  legend_breaks = NA,
  legend_labels = NA,
  show_rownames = T,
  show_colnames = T,
  fontsize = 8,
  cellwidth = 10,
  cellheight = 8,
  width = 14,
  height = 14
)
}
\arguments{
\item{width_df}{dataframe, the result of function Merge_data_for_heatmap.}

\item{filename}{character, file name saved for this pheatmap}

\item{color_pheatmap_set, }{default is NULL, it can be grDevices::colorRampPalette(c('blue','white','red'))(100), or viridis::viridis(8)}

\item{colors_list}{default is NULL, it can be colors_list <- list(species = c(Mm = rgb(102/255,46/255,115/255), Zf = rgb(31/255,153/255,139/255)))}

\item{annotation_names_col}{logical, whether annotation the column}

\item{border_color}{character. color of the border}

\item{scale}{default is none, scale by row or by column}

\item{cluster_rows}{logical, whether cluster the rows}

\item{cluster_cols}{logical, whether cluster the columns}

\item{legend}{logical, whether draw the legend}

\item{legend_breaks}{default is NA}

\item{legend_labels}{default is NA}

\item{show_rownames}{logical, whether show rownames}

\item{show_colnames}{logical, whether show colnames}

\item{fontsize}{integer, size of the font}

\item{cellwidth}{integer, size of the cell width}

\item{cellheight}{integer, size of the cell height}

\item{width}{integer, the width of the pheatmap}

\item{height}{integer, the height of the pheatmap}
}
\value{
a pheatmap in your directory
}
\description{
Plot a heatmap for average exprssion of ligand and receptor pair in two species
}
\examples{
colors_list <- list(species = c(Mm = rgb(102/255,46/255,115/255), Zf = rgb(31/255,153/255,139/255)))
pheatmap_LR(width_df = merge_width_df, filename = "lr_avg_pheatmap.pdf", colors_list = colors_list)
}
