% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheatmap_LR1.R
\name{pheatmap_LR1}
\alias{pheatmap_LR1}
\title{Create a pheatmap for average expression of ligand-receptor pairs}
\usage{
pheatmap_LR1(
  width_df,
  filename = "pheatmap_LR.pdf",
  color_pheatmap_set = NULL,
  species_colors = NULL,
  species_name1 = "mm",
  species_name2 = "zf",
  annotation_names_col = TRUE,
  border_color = "white",
  scale = "none",
  cluster_rows = FALSE,
  cluster_cols = FALSE,
  legend = TRUE,
  legend_breaks = NA,
  legend_labels = NA,
  show_rownames = T,
  show_colnames = T,
  fontsize = 10,
  cellwidth = 16,
  cellheight = 12,
  width = 14,
  height = 14,
  ...
)
}
\arguments{
\item{width_df}{width dataframe, result of function make_pheatmap_LR_data1}

\item{filename}{name to save the pheatmap, default is "pheatmap_LR.pdf"}

\item{color_pheatmap_set}{colors for the pheatmap, default is viridis::viridis(8)}

\item{species_colors}{vector, colors vectors for the species}

\item{species_name1}{two character, like "mm", for species1}

\item{species_name2}{two character, like "zf", for species2}

\item{annotation_names_col}{logical, whether show the column title name}

\item{border_color, }{character, for the color of the border, like "white"}

\item{scale}{how to scale the data, one of "none", "row", "column"}

\item{cluster_rows}{logical, whether cluster the row}

\item{cluster_cols}{logical, whether cluster the column}

\item{legend}{logical, whether show the legend}

\item{legend_breaks}{break legend with this vector value}

\item{legend_labels}{labels for the break legend}

\item{show_rownames}{logical, whether show the rownames}

\item{show_colnames}{logical, whether show the colnames}

\item{fontsize}{numeric, size of the font}

\item{cellwidth}{numeric, width of the cell}

\item{cellheight}{numeric, heigth of the cell}

\item{width}{numeric, width of the heatmap}

\item{height}{numeric, height of the heatmap}

\item{...}{other parameters in pheatmap}
}
\value{
pheatmap
}
\description{
make a pheatmap for average expression of ligand-receptor pairs using geometric mean of corresponding cell type
}
\examples{
merge_avg_width_df <- make_pheatmap_LR_data1(cci_conserved_results = conserved_result_mm_zf,
seurat_object_sp1 = Mm_seurat,
seurat_object_sp2 = Zf_seurat,
avg_group = "cellname",
species_name1 = "mm",
species_name2 = "zf",
cci_conserved_Weights = conserved_cci_result$cci_conserved_Weights)
pheatmap_LR_multi(width_df = merge_avg_width_df$merge_width_df, filename = "lr_avg_pheatmap.pdf")
}
