% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interaction_heatmap.R
\name{plot_interaction_heatmap}
\alias{plot_interaction_heatmap}
\title{Heapmap of cell-cell interaction}
\usage{
plot_interaction_heatmap(
  SingleCellSignalR_sp1_result,
  SingleCellSignalR_sp2_result,
  SingleCellSignalR_sp3_result,
  value_name = "scale_weight",
  used_col = c("source", "target", "LRscore_scale"),
  species_name1 = "mm",
  species_name2 = "zf",
  species_name3 = "ch",
  sp1_conserved_celltype,
  sp2_conserved_celltype,
  sp3_conserved_celltype,
  celltype_colors = NULL,
  brewer.pal_set = "Set3",
  species_colors = NULL,
  ph_colors = viridis::viridis(8),
  border_color = "white",
  scale = "none",
  cluster_rows = FALSE,
  cluster_cols = FALSE,
  legend = TRUE,
  show_rownames = F,
  show_colnames = F,
  fontsize = 8,
  cellwidth = 20,
  cellheight = 24,
  angle_col = "45",
  annotation_names_col = TRUE,
  annotation_names_row = TRUE,
  filename = "CCI_pheatmap.pdf",
  width = 18,
  height = 10,
  ...
)
}
\arguments{
\item{SingleCellSignalR_sp1_result}{dataframe, result of the SingCellSignalR, must contain "source", "target", "LRscore_scale"}

\item{SingleCellSignalR_sp2_result}{dataframe, result of the SingCellSignalR, must contain "source", "target", "LRscore_scale"}

\item{SingleCellSignalR_sp3_result}{dataframe, result of the SingCellSignalR, must contain "source", "target", "LRscore_scale"}

\item{value_name}{character, names of the sum weight, default is "scale_weight"}

\item{used_col}{default is c("source", "target", "LRscore_scale"), they should be in the column of result of SingCellSignalR}

\item{species_name1}{character, species_name must be only two characters, like "mm"}

\item{species_name2}{character, species_name must be only two characters, like "zf"}

\item{species_name3}{character, species_name must be only two characters, like "ch"}

\item{sp1_conserved_celltype}{vector, a vector contained conserved cell type names in species1}

\item{sp2_conserved_celltype}{vector, a vector contained conserved cell type names in species2}

\item{sp3_conserved_celltype}{vector, a vector contained conserved cell type names in species3}

\item{celltype_colors}{vector, colors vector with names. names of celltype_colors must be paste0(species_name, celltype_name), like "mmCones"}

\item{brewer.pal_set}{character, pal in RColorBrewer}

\item{species_colors}{vector, colors vector without names. colors for species}

\item{ph_colors}{vector, colors for pheatmap}

\item{border_color, }{character, colors for border}

\item{scale, }{how to scale the data, one of "none", "row", "column"}

\item{cluster_rows}{logical, whether cluster the row}

\item{cluster_cols}{logical, whether cluster the column}

\item{legend}{logical, whether show the legend}

\item{show_rownames}{logical, whether show the rownames}

\item{show_colnames}{logical, whether show the colnames}

\item{fontsize}{numeric, size of the font}

\item{cellwidth}{numeric, width of the cell}

\item{cellheight}{numeric, heigth of the cell}

\item{angle_col}{numeric, angle of the colnames}

\item{annotation_names_col}{logical, whether show the name of column}

\item{annotation_names_row}{logical, whether show the name of row}

\item{filename}{the names used to save the heatmap}

\item{width}{numeric, width of the heatmap}

\item{height}{numeric, height of the heatmap}

\item{...}{other parameters in pheatmap}
}
\value{
a list contain the data and the pheatmap object
}
\description{
Heapmap of cell-cell interaction
}
\section{Functions}{
\itemize{
\item \code{plot_interaction_heatmap()}: heapmap to show interaction between conserved cell types

}}
\examples{
mm_conserved_celltype <- c("Activated MG", "Cones", "GABAergic AC","Microglia", "Pericytes", "Resting MG", "RGC", "Rod BC", "Rods", "RPE", "V/E cells")
zf_conserved_celltype <- c("Activated MG",  "Cones", "GABAergic AC", "Microglia", "Pericytes", "Resting MG", "RGC","Cone BC", "Rods", "RPE", "V/E cells")
ch_conserved_celltype <- c("Activated MG", "Cones", "GABAergic AC", "Resting MG", "RGC", "Cone BC", "Rods")
plot_interaction_heatmap(SingleCellSignalR_sp1_result = SingleCellSignalR_mouse_result,
                         SingleCellSignalR_sp2_result = SingleCellSignalR_zebrafish_result,
                         SingleCellSignalR_sp3_result = SingleCellSignalR_chick_result,
                         species_name1 = "mm",
                         species_name2 = "zf",
                         species_name3 = "ch",
                         sp1_conserved_celltype = mm_conserved_celltype,
                         sp2_conserved_celltype = zf_conserved_celltype,
                         sp3_conserved_celltype = ch_conserved_celltype)
}
