% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Celltype.Communication.R
\name{Plot_Celltype.Communication}
\alias{Plot_Celltype.Communication}
\title{Plot communication between cell types}
\usage{
Plot_Celltype.Communication(
  graph_df,
  sendercolumn,
  receivercolumn,
  widthcolumn,
  colors = NULL,
  useLabels = TRUE,
  nodeSize = 5,
  vertex.label.color,
  edge.label.color,
  vertex.label.cex,
  ...
)
}
\arguments{
\item{graph_df}{The input dataframe containing communication data, see the example in tutorial.}

\item{sendercolumn}{single column name, The name of the column containing sender cell types.}

\item{receivercolumn}{single column name, The name of the column containing receiver cell types.}

\item{widthcolumn}{single column name, The name of the column containing the width of edges (communication strength).}

\item{colors}{Optional vector of colors for nodes and edges.}

\item{useLabels}{Logical value indicating whether to display labels for nodes.}

\item{nodeSize}{Size of the nodes.}

\item{vertex.label.color}{Color of node labels.}

\item{vertex.label.cex}{Size of node labels.}

\item{...}{}
}
\value{
The plotted graph as an invisible object
}
\description{
This function takes a dataframe containing information about communication between cell types,
and plots a graph to visualize the communication pattern. The graph represents sender and
receiver cell types as nodes, and the strength of communication as the width of edges connecting them.
}
\examples{
plot_graph_network(long_df, sendercolumn = "celltype1", receivercolumn = "celltype2", widthcolumn = "score", useLabels = T, colors = colors, nodeSize = 10, vertex.label.color = "black", vertex.label.cex = 1.5)

}
