% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Plot_MarkersExp}
\alias{Plot_MarkersExp}
\title{Title}
\usage{
Plot_MarkersExp(
  Seurat_object,
  Gene1,
  Marker_list,
  File1,
  ident.include = NULL,
  width = 12,
  height = 10,
  PlotType1 = c("ggPlot", "DotPlot", "VlnPlot", "FeaturePlot")
)
}
\arguments{
\item{Seurat_object}{Seurat object which should contain clustering inforamtion, reduction information(UMAP or TSNE)}

\item{Gene1}{character, indicating the genes you want to plot}

\item{Marker_list}{Marker gene for each cell types}

\item{ident.include}{numeric, indicating which clusters include in the plot}

\item{width}{numeric, indicating width of the figure}

\item{height}{numeric, indicating height of the figure}

\item{PlotType1}{types of plot, including: 'ggplot', 'Dotplot', 'VlnPlot' and 'FeaturePlot'}

\item{fileprefix}{the prefix of output figure}
}
\value{

}
\description{
Title
}
