% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MarkersHeatmap.R
\name{plot_MarkersHeatmap}
\alias{plot_MarkersHeatmap}
\title{Title}
\usage{
plot_MarkersHeatmap(scRNA1, Color1 = NULL, ModuleColor1 = NULL, Gene1 = NULL)
}
\arguments{
\item{scRNA1}{dataframe, generated by Format_Markers_Frac}

\item{Color1}{vector of colors used in heatmap}

\item{ModuleColor1}{colors for each moudle}

\item{Gene1}{genes that you want to show in the pheatmap}
}
\value{

}
\description{
Title
}
