% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Species_CellType_Tree.R
\name{Plot_Species_CellType_Tree}
\alias{Plot_Species_CellType_Tree}
\title{Plot Phylogenetic Tree of Different Species Cell Types}
\usage{
Plot_Species_CellType_Tree(
  dist_matrix,
  hcluster.method = c("average", "single", "complete"),
  species.vector,
  layout.tree = c("rectangular", "dendrogram", "fan", "circular"),
  offset = 0.01,
  tiplab.size = 5,
  tippoint.shape = c(22, 21),
  tippoint.shape.size = 4,
  geom_nodepoint = 0,
  col.value = c(rgb(102/255, 46/255, 115/255), rgb(31/255, 153/255, 139/255))
)
}
\arguments{
\item{dist_matrix}{Distance matrix}

\item{hcluster.method}{chariter, Hierarchical clustering method, possible values are "average" (default), "single", "complete"}

\item{species.vector}{vector, Species vector}

\item{layout.tree}{Tree layout ("rectangular", "dendrogram", "fan", "circular"), default is "rectangular"}

\item{offset}{Tiplab offset, default is 0.01}

\item{tiplab.size}{Tiplab size, default is 5}

\item{tippoint.shape}{Tippoint shape,  default is 22}

\item{tippoint.shape.size}{Tippoint size, default is 4}

\item{geom_nodepoint}{Nodepoint size, default is 0}

\item{col.value}{Color vector for species, default is c(rgb(102/255,46/255,115/255), rgb(31/255,153/255,139/255))}
}
\value{
a list, which contain the Tree data and the Tree plot
}
\description{
Plot a tree of different species cell types using the conserved score of cell types
}
\examples{
load(system.file("extdata", "zf_mm_markers.rda", package = "CACIMAR"))
expression <- Identify_ConservedCellTypes(OrthG_Mm_Zf,Zf_marker,Mm_marker,'zf','mm')
dist_matrix <- expression[[2]]
species.vector <- substr(rownames(dist_matrix), 1, 2)
Species_CellType_Tree <- Plot_Species_CellType_Tree(dist_matrix = dist_matrix, species.vector = species.vector, hcluster.method = "average", geom_nodepoint = 0, layout.tree = "rectangular", offset = 0.02, tiplab.size = 5, tippoint.shape = 22, tippoint.shape.size = 4)
Species_CellType_Tree$Plot
}
