% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refine_markers.R
\name{Refine_Markers}
\alias{Refine_Markers}
\title{inner function to refine markers}
\usage{
Refine_Markers(Seurat_obejct, Spec1, GeneSymb1, Marker, FracThr1 = 3)
}
\arguments{
\item{Seurat_obejct}{seurat object}

\item{Spec1}{character, indicating the species of data, including: 'Mm', 'Ch', 'Zf', 'Hs'}

\item{GeneSymb1}{Gene correspondence that is used for ID chaning. first column should be ENSEMBLE ID, second column should be Symbol, third column should be NCBIID and fourth column should be officalsymbol.}

\item{Marker}{Marker gene list}

\item{FracThr1}{????}
}
\value{

}
\description{
inner function to refine markers
}
