% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-species_markers.R
\name{Refine_Markers_Species}
\alias{Refine_Markers_Species}
\title{Inner function to refine markers}
\usage{
Refine_Markers_Species(Marker1, Species)
}
\description{
Inner function to refine markers
}
