% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-species_markers.R
\name{Refine_TwoSpecies}
\alias{Refine_TwoSpecies}
\title{Refine cluster markers between two species}
\usage{
Refine_TwoSpecies(ShMarker11, CellT11, CellT12, Species)
}
\arguments{
\item{Species}{}
}
\value{

}
\description{
Refine cluster markers between two species
}
