% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Rename_CellType}
\alias{Rename_CellType}
\title{Title}
\usage{
Rename_CellType(RNA1, CellType)
}
\arguments{
\item{RNA1}{correlation of expression in each cell type}

\item{CellType}{cell type. First column should be Species, second column should
be Cluster, third column should be CellType, fourth column should be Group}
}
\value{

}
\description{
Title
}
