% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_MarkersExp.R
\name{Seurat_SubsetData}
\alias{Seurat_SubsetData}
\title{Inner function to subset seurat object}
\usage{
Seurat_SubsetData(pbmc1, SubG1, SubS1 = NULL, ExSubS1 = NULL)
}
\description{
Inner function to subset seurat object
}
