% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChordDiagram.R
\name{show_LR_pairs}
\alias{show_LR_pairs}
\title{Plot ligand-receptor pairs in a dot plot}
\usage{
show_LR_pairs(
  LR_data,
  fill_colors = NULL,
  brewer.pal_set = "Paired",
  color_x = "black",
  color_x_in_group = F,
  color_y = "black",
  size = 9,
  shape = 21,
  colour = "white",
  stroke = 2,
  theme_set = theme_bw(),
  text_size = 40,
  text_colour = "black",
  legend.position = "none"
)
}
\arguments{
\item{LR_data}{data frame only containing 3 column representing the ligand-receptor pairs, cell type pairs, and groups}

\item{fill_colors}{vector, colors for the points}

\item{brewer.pal_set}{if fill_colors = NULL, this can set the brewer.pal for colors used for the text and points. By default, brewer.pal_set = "Paired"}

\item{color_x}{By default, colors for x text is set to be "black"}

\item{color_x_in_group}{if color_x_in_group = T, points colors will be used for x text}

\item{color_y}{By default, colors for y text is set to be "black", or it can be set to any colors you want, like "red", "blue"}

\item{size}{numeric, size for points}

\item{shape}{numeric, shape for points}

\item{colour}{colors for points}

\item{stroke}{numeric, width of the point’s outline}

\item{theme_set}{theme used. By default it is theme_bw()}

\item{text_size}{numeric, size of the text in this plot}

\item{text_colour}{colors for the text in this plot}

\item{legend.position}{the position of the legend. By default, we don't show the legend}
}
\value{
A dot plot. Row is the ligand-receptor pairs, while column is the cell type pairs
}
\description{
Construct a dot plot to show the conserved ligand-receptor pairs
}
\examples{
show_LR_pairs(LR_data = LRpair_show_0.75,  color_x_in_group = T)

}
