% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cal_SinlgeCell_Cor_SmoothByBin.R
\name{smoothByBin}
\alias{smoothByBin}
\title{Internal function}
\usage{
smoothByBin(
  Exp1,
  Pseudotime1,
  PseudotimeRange1 = NULL,
  SmoothLength1 = 100,
  ByBin1 = c("Equal.Pseudotime", "Equal.Cells")
)
}
\description{
Internal function
}
