% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_Cell_interaction_sankey.R
\name{create_sankey}
\alias{create_sankey}
\title{Create sankey plot}
\usage{
create_sankey(
  links,
  specie_name1 = "Mm",
  specie_name2 = "Zf",
  output_file = "sankey.html",
  colors_file = NULL,
  brewer.pal_set = "Set3",
  node_width = 18,
  node_padding = 15,
  node_stroke_width = 0,
  node_corner_radius = 0,
  drag_x = TRUE,
  drag_y = TRUE,
  units = "TWh",
  node_pos_x = "group",
  align = "none",
  scale_node_breadths_by_string = F,
  show_node_values = FALSE,
  height = 1400,
  width = 1800,
  linkColor = "#A0A0A0",
  link_type = "bezier",
  curvature = 0.5,
  link_opacity = 5,
  link_gradient = FALSE,
  node_shadow = FALSE,
  node_label_margin = 5,
  zoom = TRUE,
  font_family = NULL,
  iterations = 0,
  x_scaling_factor = 1.4,
  font_size = 16,
  ...
)
}
\arguments{
\item{links}{dataframe, a links file containing source, target, and weight}

\item{specie_name1}{character, name for specie 1}

\item{specie_name2}{character, name for specie 2}

\item{output_file}{character, save sankey plot with this file name}

\item{colors_file}{character, file path for the color file, it is NULL by default and it will assign colors automatically}

\item{brewer.pal_set}{if colors_file is null, the brewer.pal can be choose, like "Set1", "Set3", "Paired", ......}

\item{node_width}{numeric, width of the node}

\item{node_padding}{numeric, distance between the node}

\item{node_stroke_width}{numeric, width of the stroke around nodes}

\item{node_corner_radius}{numeric, the radius of the node}

\item{drag_x}{logical, TRUE indicates that the plot can be horizontally dragged}

\item{drag_y}{logical, TRUE indicates that the plot can be vertically dragged}

\item{units}{character, name for units}

\item{node_pos_x}{character, variable used for grouping nodes on the x-axis}

\item{align}{character, alignment of the nodes. One of 'right', 'left', 'justify', 'center', 'none'. If 'none', then the labels of the nodes are always to the right of the node}

\item{scale_node_breadths_by_string}{logical, Put nodes at positions relatively to string lengths - only work well currently with align='none'}

\item{show_node_values}{logical, Show values above nodes. Might require and increased node margin.}

\item{height}{numeric, height of the sankey plot for saving}

\item{width}{numeric, width of the sankey plot for saving}

\item{linkColor}{numeric, color of links}

\item{link_type}{character, one of 'bezier', 'l-bezier', 'trapezoid', 'path1' and 'path2'}

\item{curvature}{numeric, curvature parameter for bezier links - between 0 and 1}

\item{link_opacity}{numeric, opacity of links}

\item{link_gradient}{logical, add a gradient to the links}

\item{node_shadow}{logical, add a shadow to the nodes}

\item{node_label_margin}{numeric, distance between the node and font}

\item{zoom}{logical, value to enable (TRUE) or disable (FALSE) zooming}

\item{iterations}{numeric, number of iterations in the diagramm layout for computation of the depth (y-position) of each node. Note: this runs in the browser on the client so don't push it too high}

\item{x_scaling_factor}{numeric, scale the computed x position of the nodes by this value}

\item{font_size}{numeric, size of the font of the plot}

\item{...}{param of sankeyD3::sankeyNetwork()}
}
\value{
a sankey plot
}
\description{
Build a sankey plot to show the cell-cell interaction profile of two species
}
\examples{
create_sankey(links = all_weight_df_long[, c("Source2", "target", "scale_weight")],
output_file = "sankey_scale_weight.html",
specie_name1 = "Mm",
specie_name2 = "Zf",
colors_file = NULL)

}
