% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Heatmap_Cor}
\alias{Heatmap_Cor}
\title{plot the heatmap of marker genes across different species}
\usage{
Heatmap_Cor(
  RNA1,
  CellType,
  RowType1 = "",
  ColType1 = "",
  cluster_cols = T,
  cluster_rows = F,
  Color1 = NULL
)
}
\arguments{
\item{RNA1}{correlation of expression in each cell type}

\item{CellType}{cell type. First column should be Species, second column should
be Cluster, third column should be CellType, fourth column should be Group}

\item{RowType1}{character, indicating the cell types that you want to show
on the row in heatmap. RowType1='' means show all cell types}

\item{ColType1}{character, indicating the cell types that you want to show
on the column in heatmap. RowType1='' means show all cell types}

\item{cluster_cols}{boolean values determining if columns should be clustered
or hclust object}

\item{cluster_rows}{boolean values determining if rows should be clustered or
hclust object}

\item{Color1}{vector of colors used in heatmap}
}
\value{
return a list used in Plot_tree
}
\description{
plot the heatmap of marker genes across different species
}
