% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_Markers.R
\name{Identify_Markers}
\alias{Identify_Markers}
\title{Identify markers of each cluster}
\usage{
Identify_Markers(Seurat_object, RocThr = 0.4, DiffThr = 1/3, FracThr = 3)
}
\arguments{
\item{Seurat_object}{Seurat object, should contain cluster information}

\item{RocThr}{numeric, indicating the cutoff of Roc to identify marker genes}

\item{DiffThr}{numeric, indicating the cutoff of difference in marker genes between
clusters}

\item{FracThr}{numeric, indicating the threshold of estimate of the odds
ratio}
}
\value{
Data frame of conserved markers
}
\description{
This function first identify marker genes in each cluster
with Roc threshold > RocThr. Then, based on marker genes identified above,
this function calculates the difference and power of marker genes in each
cluster, and marker genes with Difference threshold > DiffThr will be retained.
Next, gene with the largest power in which cluster will be the marker
gene in this cluster. Eventually, make fisher test for power of each cluster,
cluster with p.value < 0.05 will be retained as the final cluster for marker gene
}
