% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypeFraction.R
\name{Identify_SharedMarkers}
\alias{Identify_SharedMarkers}
\title{Calculate fraction of shared markers between species}
\usage{
Identify_SharedMarkers(Marker_list, Species_names, PowerTh1 = 0.1)
}
\arguments{
\item{PowerTh1}{}
}
\value{

}
\description{
Calculate fraction of shared markers between species
}
