% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-species_markers.R
\name{OrthG_ThreeSpecies}
\alias{OrthG_ThreeSpecies}
\title{Identify orthologs genes for three species}
\usage{
OrthG_ThreeSpecies(
  OrthG_TwoSpecies1,
  OrthG_TwoSpecies2,
  Species_name,
  match_cell_name
)
}
\arguments{
\item{OrthG_TwoSpecies1}{First two species orthologs marker genes,
generated by \code{\link{OrthG_Twospecies}}. One species in OrthG_TwoSpecies1
should overlap with one species in OrthG_TwoSpecies2}

\item{OrthG_TwoSpecies2}{Second two species orthologs marker genes,
generated by \code{\link{OrthG_Twospecies}}. One species in OrthG_TwoSpecies2
should overlap with one species in OrthG_TwoSpecies1}

\item{Species_name}{character, indicating the names of species in OrthG_TwoSpecies1
and OrthG_TwoSpecies2}

\item{match_cell_name}{characters contained in both cell names
to match similar cell types}
}
\value{

}
\description{
Identify orthologs genes for three species based on the result
of \code{\link{OrthG_Twospecies}}
}
