% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_MarkersExp.R
\name{plot_MarkersExp}
\alias{plot_MarkersExp}
\title{Inner function to plot MarkersExp}
\usage{
plot_MarkersExp(
  pbmc1,
  File1,
  Marker_list,
  PlotType1 = c("FeaturePlot", "DotPlot", "VlnPlot", "ggPlot"),
  Gene1 = "",
  CellType1 = "",
  ident.include = NULL,
  group.by = NULL,
  width = 12,
  height = 10,
  nrow = 3,
  RedMeth1 = "tSNE",
  no.legend = F,
  legend.text.size = 12,
  legend.label = T,
  size.title.use = 10,
  size.axis.text = 10,
  point.size.use = 2,
  Xtick1 = T,
  dot.scale = 10,
  IDtype1 = c("EnsID", "Symbol"),
  Col2 = NULL
)
}
\description{
Inner function to plot MarkersExp
}
