% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Plot_MarkersHeatmap}
\alias{Plot_MarkersHeatmap}
\title{Plot the Marker genes in each cluster}
\usage{
Plot_MarkersHeatmap(
  scRNA1,
  ModuleScale1 = 15,
  Color1 = NULL,
  ModuleColor1 = NULL,
  Gene1 = NULL,
  NumRowBlank = 0,
  show_colnames = F,
  legend1 = T,
  Scale = "none"
)
}
\arguments{
\item{scRNA1}{dataframe, generated by}

\item{ModuleScale1}{change the relative proportion of module bar and expression heatmap}

\item{Color1}{vector of colors used in heatmap}

\item{ModuleColor1}{colors for each moudle}

\item{Gene1}{genes that you want to show in the pheatmap}

\item{NumRowBlank}{the blank between each group}

\item{show_colnames}{logic, indicating whether show column names}

\item{legend1}{logic, indicating whether show the legend}

\item{Scale}{character indicating if the values should be centered and scaled
in either the row direction or the column direction, or none. Corresponding
values are "row", "column" and "none"}
}
\value{
heatmap
}
\description{
Plot the Marker genes in each cluster
}
\examples{
data("pbmc_small")
all.markers <- Identify_Markers(pbmc_small,,RocThr = 0)
Plot_MarkersHeatmap(all.markers)
}
