% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phylogenetic_tree.R
\name{Plot_phylogenetic_tree}
\alias{Plot_phylogenetic_tree}
\title{Plot phylogenetic tree}
\usage{
Plot_phylogenetic_tree(
  SCT_matrix,
  tree_method = "hierarchical clustering",
  hcluster.method = "average",
  species.vector,
  layout.tree = "rectangular",
  tree_size = 1.4,
  xlim_tree = 0.65,
  fontface_for_tiplab = TRUE,
  set_fontface_for_tiplab = c("bold.italic", "italic", "plain"),
  conserved_hm_celltype = NULL,
  bar_width = 0.025,
  Show_Pairwise_distances = FALSE,
  boot_times = 100,
  offset = 0.04,
  offset2 = 0.022,
  offset_tiplab = 0.07,
  fontface_tiplab_with_colors = "bold",
  tiplab.size = 7,
  tiplab_cols = NULL,
  tippoint.shape = 21,
  tippoint.shape.size = 0,
  geom_nodepoint = 5,
  show_branch.length = FALSE,
  round_x = 2,
  fill_brandlength = "lightgreen",
  size_brandlength = 6,
  annotation_colors_df = NULL,
  brewer_pal_used = "Set3",
  col.value = c(rgb(102/255, 46/255, 115/255), rgb(31/255, 153/255, 139/255)),
  colors_labels = NULL,
  show_colnames = FALSE,
  colnames_angle = 0,
  colnames_position = "top",
  font.size_colnames = 5,
  colnames_offset_y = 0.8,
  custom_column_labels = c("Species", "Celltypes"),
  bootstrap_value_size = 0,
  bootstrap_value_col = "black",
  legend.text_size = 18,
  width = 13,
  height = 22,
  plot.margin = margin(10, 10, 10, -10),
  legend.position = "top"
)
}
\arguments{
\item{SCT_matrix}{matrix, contain the conservation scores of cell types across species}

\item{tree_method}{character, method used for constructing phylogenetic tree, it can be "hierarchical clustering", or "classic Neighbor-Joining"}

\item{hcluster.method}{character, method used for clustering phylogenetic tree, it can be one of "average", "single", "complete", "mcquitty", "median", "centroid"}

\item{species.vector}{vector, length is equal to nrow(SCT_matrix), used for annotation bar}

\item{layout.tree}{character, method used for the layout of the tree. it can be one of 'rectangular', 'dendrogram', 'fan', 'circular'}

\item{tree_size}{numeric, size of the tree's branch}

\item{xlim_tree}{numeric, x limit for the picture, used for the width of the picture}

\item{fontface_for_tiplab}{logical, whether used fontface for tip label. If it is TRUE, then the tiplab_cols for tip label won't work}

\item{set_fontface_for_tiplab}{if fontface_for_tiplab is TRUE, the parameter will work. By default, it is set with c("bold.italic", "italic", "plain")}

\item{conserved_hm_celltype}{vector, contain the conserved cell types in heatmap, or celltypes defined by yourself}

\item{bar_width}{integer, width of annotation bar}

\item{Show_Pairwise_distances}{logical, whether save the plot of pairwise distances between input data and constructed tree or not}

\item{boot_times}{integer, the times for bootstrap test, usually 100, or you can set larger 1000}

\item{offset}{numeric, the distance of celltype annotation bar from tip}

\item{offset2}{numeric, the distance of species annotation bar from tip}

\item{offset_tiplab}{numeric, the distance of labels from tip}

\item{fontface_tiplab_with_colors}{charcter, colors for tip label. if fontface_for_tiplab is FALSE, then can set one fontface for all tip labels, like "bold", "italic", "plain"}

\item{tiplab.size}{numeric, the size of tip labels}

\item{tiplab_cols}{vector, colors for tip group}

\item{tippoint.shape}{numeric, the shape used for tip point, such as 21, 17, 22, ......}

\item{tippoint.shape.size}{numeric, the size of the tip point shape}

\item{geom_nodepoint}{integer, the size of the node point}

\item{show_branch.length}{logical, whether show the branch length, default is TRUE}

\item{round_x}{numeric, the decimal places contained for round() function for branch length number}

\item{fill_brandlength}{character, colors for branch length}

\item{size_brandlength}{numeric, the size for the branch length label}

\item{annotation_colors_df}{dataframe, must contain two column, named "colors" and "celltype", default is null}

\item{brewer_pal_used}{character, the brewer_pal in RColorBrewer, like "Set1",  "Set3", "Paired". It set colors for tip label. It works when fontface_for_tiplab is FALSE}

\item{col.value}{chacterize vector, colors used for species bar annotation, must be the same length of the number of species}

\item{colors_labels}{chacterize vector, labels for the legend, the length of colors_breaks}

\item{show_colnames}{logical, whether show the colnames of annotation bar}

\item{colnames_angle}{integer, the angle of the colnames of annotation bar}

\item{colnames_position}{character, the position of the colnames of annotation bar, can be top or bottom}

\item{font.size_colnames}{integer, the size of the colnames of annotation bar}

\item{colnames_offset_y}{numeric, the distance of the colnames from the annotation bar}

\item{custom_column_labels}{character, edit the colnames of annotation bar by yourself}

\item{bootstrap_value_size}{numeric, the size of the bootstrap values showed in the tree. If you don't want to showed the  bootstrap values in the tree, you can set the parameter to 0}

\item{bootstrap_value_col}{character, the color of the bootstrap values showed in the tree}

\item{legend.text_size}{integer, the size of the legend}

\item{width}{numeric, the width of the plot}

\item{height}{numeric, the height of the plot}

\item{plot.margin}{the plot margin, it should be something like margin(10, 10, 10, -10), the four number represent top, right, bottom, left, respectively}

\item{legend.position}{character, one of "right", "left", "top", "bottom", the position of the legend}

\item{fontface_tiplab}{charcter, the fontface used for tip label}
}
\value{
save plot in present file directory
}
\description{
Build a phylogenetic tree to identify the conservation of the cell types across species
}
\examples{
library(CACIMAR, lib.loc = "/usr/local/lib/R/site-library")
load(system.file("extdata", "zf_mm_markers.rda", package = "CACIMAR"))
expression <- CACIMAR::Identify_ConservedCellTypes(OrthG_Mm_Zf,Zf_marker,Mm_marker,'zf','mm')
SCT_matrix <- expression[[2]]
# get the species names
species.vector <- substr(rownames(SCT_matrix), 1, 2)
# get the conserved celltypes in heatmap
SNT_h <- SCT_matrix[grep('mm',rownames(SCT_matrix)),as.numeric(grep('zf',colnames(SCT_matrix)))]
conserved_hm_celltypes <- get_conserved_hm_celltypes(SNT_h)
Plot_phylogenetic_tree(SCT_matrix = SCT_matrix,
                       species.vector = species.vector,
                       conserved_hm_celltype = conserved_hm_celltypes)

}
