% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sort_MarkersPower.R
\name{Sort_MarkersPower}
\alias{Sort_MarkersPower}
\title{Get cluster with the maximal power for each marker genes}
\usage{
Sort_MarkersPower(MarkerRoc01, Thr01 = 0)
}
\arguments{
\item{MarkerRoc01}{markerroc01}

\item{Thr01}{thr01}
}
\value{
?
}
\description{
Get cluster with the maximal power for each marker genes
}
